Remove the Registrations and Sessions controllers
These controllers were introduced in order to change the redirection
behaviour of the parent Devise controllers. After reading up on the
source code of Devise it became clear that these changes of behaviour
could be done with the introduction of a single route -- the user_root
route.

This commit introduces the user_root route, removes the redundand
controllers and modifies the specs in order to continue passing. Note:
These specs would be better off as integration tests.