Display only reachable neighbours
After the kernel update on cassie, the presence system seemed to display
people that were long gone from the lab. It seems that the timings on
the entry expiration in the neighbour table has changed with the new
kernel version. As suggested by @krokodilerian, use the ip neigh show
reachable command in order to only list the reachable neighbours.